feat(summary): differentiate details based on media type
